Use submodules to organize dashboard widgets, list filters and paginator
